Fix float slow and ulti

+++ Classes/Update.cpp
@@ -169,7 +169,7 @@ float Update::getStunTime()
 	return this->stunTime;
 }
 
-int Update::getConditionUlti()
+float Update::getConditionUlti()
 {
 	return this->conditionUlti;
 }
